ml/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
clip==1.0
filelock==3.12.4
fsspec==2023.10.0
ftfy==6.1.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2
numpy==1.26.1
Pillow==10.1.0
regex==2023.10.3
requests==2.31.0
sympy==1.12
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.8
